refactor(bezier-codemod): separate test util function
